+++ gulpfile.babel.js
@@ -314,6 +314,8 @@ npm_copy_task("rangy/lib/**", "external/rangy");
 
 npm_copy_task("bootstrap-notify/bootstrap-notify*.js");
 
+npm_copy_task("typeahead.js-bootstrap-css/typeaheadjs.css");
+
 gulp.task("build-info", Promise.coroutine(function* () {
     const dest = "build/standalone/lib/wed/build-info.js";
     const is_newer = yield newer(["lib/**", "!**/*_flymake.*"], dest);
@@ -344,7 +346,6 @@ gulp.task("build-standalone",
               "build-only-standalone-less",
               "build-only-standalone-config",
               "copy-log4javascript",
-              "copy-typeaheadjs.css",
               copy_tasks,
               "build-schemas",
               "build-samples",